LayoutTests/platform/mac/editing/deleting/id-in-deletebutton-expected.txt

Test document.getElementById("WebKit-Editing-Delete-Button")

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.findEditingDeleteButton(); document.getElementById("WebKit-Editing-Delete-Button") is non-null.
PASS successfullyParsed is true

TEST COMPLETE
1
2
